requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
pydantic>=2.5.0
pydantic-settings>=2.1.0
azure-ai-documentintelligence>=1.0.0b1
azure-search-documents>=11.4.0
azure-core>=1.29.5
openai>=1.26.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
langchain-text-splitters>=0.2.0
langchain-community>=0.3.0
langchain-classic>=0.1.0
pytest>=7.4.3
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
azure-ai-projects>=1.0.0
azure-identity>=1.15.0